+++ griptape/drivers/prompt/azure_openai_chat_prompt_driver.py
@@ -29,14 +29,14 @@ class AzureOpenAiChatPromptDriver(OpenAiChatPromptDriver):
         default=Factory(lambda self: self.model, takes_self=True),
         metadata={"serializable": True},
     )
-    azure_endpoint: str = field(kw_only=True, metadata={"serializable": True})
+    azure_endpoint: Optional[str] = field(kw_only=True, default=None, metadata={"serializable": True})
     azure_ad_token: Optional[str] = field(kw_only=True, default=None, metadata={"serializable": False})
     azure_ad_token_provider: Optional[Callable[[], str]] = field(
         kw_only=True,
         default=None,
         metadata={"serializable": False},
     )
-    api_version: str = field(default="2023-05-15", kw_only=True, metadata={"serializable": True})
+    api_version: Optional[str] = field(default="2023-05-15", kw_only=True, metadata={"serializable": True})
     client: openai.AzureOpenAI = field(
         default=Factory(
             lambda self: openai.AzureOpenAI(