+++ griptape/drivers/prompt/ollama_prompt_driver.py
@@ -33,7 +33,6 @@ class OllamaPromptDriver(BasePromptDriver):
         model: Model name.
     """
 
-    model: str = field(kw_only=True, metadata={"serializable": True})
     host: Optional[str] = field(default=None, kw_only=True, metadata={"serializable": True})
     client: Client = field(
         default=Factory(lambda self: import_optional_dependency("ollama").Client(host=self.host), takes_self=True),
@@ -46,17 +45,6 @@ class OllamaPromptDriver(BasePromptDriver):
                 max_input_tokens=2000,
                 max_output_tokens=self.max_tokens,
             ),
-            takes_self=True,
-        ),
-        kw_only=True,
-    )
-    options: dict = field(
-        default=Factory(
-            lambda self: {
-                "temperature": self.temperature,
-                "stop": self.tokenizer.stop_sequences,
-                "num_predict": self.max_tokens,
-            },
             takes_self=True,
         ),
         kw_only=True,
@@ -87,7 +75,17 @@ class OllamaPromptDriver(BasePromptDriver):
     def _base_params(self, prompt_stack: PromptStack) -> dict:
         messages = self._prompt_stack_to_messages(prompt_stack)
 
-        return {"messages": messages, "model": self.model, "options": self.options}
+        return {
+            "messages": messages,
+            "model": self.model,
+            "options": {
+                "num_predict": self.max_tokens,
+                "stop": self.tokenizer.stop_sequences,
+                **({"temperature": self.temperature} if self.temperature is not None else {}),
+                **({"top_p": self.top_p} if self.top_p is not None else {}),
+                **self.additional_params,
+            },
+        }
 
     def _prompt_stack_to_messages(self, prompt_stack: PromptStack) -> list[dict]:
         return [