+++ tests/unit/drivers/prompt/test_google_prompt_driver.py
@@ -151,7 +151,7 @@ class TestGooglePromptDriver:
     def test_try_run(self, mock_generative_model, prompt_stack, messages, use_native_tools):
         # Given
         driver = GooglePromptDriver(
-            model="gemini-pro", api_key="api-key", top_p=0.5, top_k=50, use_native_tools=use_native_tools
+            model="gemini-pro", api_key="api-key", top_p=None, use_native_tools=use_native_tools
         )
 
         # When
@@ -167,7 +167,7 @@ class TestGooglePromptDriver:
         call_args = mock_generative_model.return_value.generate_content.call_args
         assert messages == call_args.args[0]
         generation_config = call_args.kwargs["generation_config"]
-        assert generation_config == GenerationConfig(temperature=0.1, top_p=0.5, top_k=50, stop_sequences=[])
+        assert generation_config == GenerationConfig(temperature=0.1, top_p=None, stop_sequences=[])
         if use_native_tools:
             tool_declarations = call_args.kwargs["tools"]
             assert [
@@ -188,7 +188,7 @@ class TestGooglePromptDriver:
     def test_try_stream(self, mock_stream_generative_model, prompt_stack, messages, use_native_tools):
         # Given
         driver = GooglePromptDriver(
-            model="gemini-pro", api_key="api-key", stream=True, top_p=0.5, top_k=50, use_native_tools=use_native_tools
+            model="gemini-pro", api_key="api-key", stream=True, top_p=None, use_native_tools=use_native_tools
         )
 
         # When
@@ -206,9 +206,7 @@ class TestGooglePromptDriver:
 
         assert messages == call_args.args[0]
         assert call_args.kwargs["stream"] is True
-        assert call_args.kwargs["generation_config"] == GenerationConfig(
-            temperature=0.1, top_p=0.5, top_k=50, stop_sequences=[]
-        )
+        assert call_args.kwargs["generation_config"] == GenerationConfig(temperature=0.1, top_p=None, stop_sequences=[])
         if use_native_tools:
             tool_declarations = call_args.kwargs["tools"]
             assert [